Create feral module
