+++ docs/lambda/directory.conf
@@ -1,5 +1,4 @@
-laika.title = Lambda
+laika.title = Feral
 laika.navigationOrder = [
     getting-started.md
-    circe.md
 ]
