+++ feral/src/main/scala/kinesis4cats/consumer/feral/KinesisStreamEvent.scala
@@ -14,19 +14,41 @@
  * limitations under the License.
  */
 
-package kinesis4cats.consumer.lambda
-package instances
+package kinesis4cats.consumer
+package feral
 
 import scala.util.Try
 
 import java.time.Instant
 
+import _root_.feral.lambda.KernelSource
 import io.circe.Decoder
 import io.circe.scodec._
+import scodec.bits.ByteVector
 
+import kinesis4cats.models.EncryptionType
 import kinesis4cats.models.instances.circe._
 
-object circe {
+final case class KinesisStreamRecordPayload(
+    approximateArrivalTimestamp: Instant,
+    data: ByteVector,
+    kinesisSchemaVersion: String,
+    partitionKey: String,
+    encryptionType: Option[EncryptionType],
+    sequenceNumber: String
+) {
+  def asRecord = Record(
+    sequenceNumber,
+    approximateArrivalTimestamp,
+    data,
+    partitionKey,
+    encryptionType,
+    None,
+    None
+  )
+}
+
+object KinesisStreamRecordPayload {
   implicit private val instantCirceDecoder: Decoder[Instant] =
     Decoder.decodeBigDecimal.emapTry { millis =>
       def round(x: BigDecimal) = x.setScale(0, BigDecimal.RoundingMode.DOWN)
@@ -47,7 +69,20 @@ object circe {
       "encryptionType",
       "sequenceNumber"
     )(KinesisStreamRecordPayload.apply)
+}
 
+final case class KinesisStreamRecord(
+    awsRegion: String,
+    eventID: String,
+    eventName: String,
+    eventSource: String,
+    eventSourceArn: String,
+    eventVersion: String,
+    invokeIdentityArn: String,
+    kinesis: KinesisStreamRecordPayload
+)
+
+object KinesisStreamRecord {
   implicit val kinesisStreamRecordCirceDecoder: Decoder[KinesisStreamRecord] =
     Decoder.forProduct8(
       "awsRegion",
@@ -59,7 +94,19 @@ object circe {
       "invokeIdentityArn",
       "kinesis"
     )(KinesisStreamRecord.apply)
+}
 
+final case class KinesisStreamEvent(
+    records: List[KinesisStreamRecord]
+) {
+  def deaggregate: Try[List[Record]] =
+    Record.deaggregate(records.map(_.kinesis.asRecord))
+}
+
+object KinesisStreamEvent {
   implicit val kinesisStreamEventCirceDecoder: Decoder[KinesisStreamEvent] =
     Decoder.forProduct1("Records")(KinesisStreamEvent.apply)
+
+  implicit def kinesisStreamEventKernelSource
+      : KernelSource[KinesisStreamEvent] = KernelSource.emptyKernelSource
 }
